Fix: Added assets section. Added more dir checks. Lots of other fixes.

+++ scripts/dist.js
@@ -1,12 +1,21 @@
 const fs = require('fs-extra');
 const { paths } = require('../constants');
 
-// Create versioned assets for package
-fs.copySync(paths.build.root, paths.dist.root);
-fs.copySync(paths.build.root, paths.dist.version);
+(async () => {
+    // Make sure build directory exists
+    if (!(await fs.pathExists(paths.build.root))) {
+        return console.error(
+            `\n[dist] ERROR: No build directory found at "${paths.build.root}"\n`
+        );
+    }
 
-// Remove unneeded files
-['package.json', 'README.md'].forEach((file) => {
-    fs.removeSync(`${paths.dist.root}${file}`);
-    fs.removeSync(`${paths.dist.version}${file}`);
-});
+    // Create versioned assets for package
+    fs.copySync(paths.build.root, paths.dist.root);
+    fs.copySync(paths.build.root, paths.dist.version);
+
+    // Remove unneeded files
+    ['package.json', 'README.md'].forEach((file) => {
+        fs.removeSync(`${paths.dist.root}${file}`);
+        fs.removeSync(`${paths.dist.version}${file}`);
+    });
+})();
